+++ src/manifest.ts
@@ -2,6 +2,7 @@ import { resolve } from 'path'
 import { createHash } from 'crypto'
 import { readJSON, remove, existsSync, readFile, writeFile, mkdirp } from 'fs-extra'
 import { ViteBuildContext } from './types'
+import { uniq, isJS, isCSS } from './utils'
 
 const DEFAULT_APP_TEMPLATE = `
 <!DOCTYPE html>
@@ -176,27 +177,6 @@ function hash (input: string, length = 8) {
     .substr(0, length)
 }
 
-function uniq<T> (arr: T[]): T[] {
-  return Array.from(new Set(arr))
-}
-
-// Copied from vue-bundle-renderer utils
-const IS_JS_RE = /\.[cm]?js(\?[^.]+)?$/
-const IS_MODULE_RE = /\.mjs(\?[^.]+)?$/
-const HAS_EXT_RE = /[^./]+\.[^./]+$/
-const IS_CSS_RE = /\.css(\?[^.]+)?$/
-
-export function isJS (file: string) {
-  return IS_JS_RE.test(file) || !HAS_EXT_RE.test(file)
-}
-
-export function isModule (file: string) {
-  return IS_MODULE_RE.test(file) || !HAS_EXT_RE.test(file)
-}
-
-export function isCSS (file: string) {
-  return IS_CSS_RE.test(file)
-}
 function getModuleIds ([, value]: [string, any]) {
   if (!value) { return [] }
   // Only include legacy and css ids
